factors out error handling
