+++ parser/src/lib.rs
@@ -3,6 +3,7 @@ extern crate lalrpop_util;
 extern crate num_rational;
 extern crate regex;
 pub mod ast;
+pub mod error_handling;
 pub mod float_to_rational;
 pub mod imports;
 pub mod parser;
