+++ parser/src/parser.rs
@@ -3,9 +3,9 @@ extern crate colored;
 extern crate num_rational;
 use crate::ast::*;
 use crate::imports::{get_filepath, get_import_name, is_import};
+use crate::error_handling::handle_parse_error;
 use colored::*;
 use num_rational::Rational64;
-use std::cmp;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::prelude::*;
@@ -84,56 +84,9 @@ pub fn parse_file(filename: &String, parse_table: Option<ParseTable>) -> ParsedC
     match init.clone() {
         Ok(init) => ParsedComposition { init, table },
         Err(error) => {
-            let start_offset = 125;
-            let end_offset = 50;
             let location = Arc::new(Mutex::new(Vec::new()));
-            let cmp_len = &composition.len();
             error.map_location(|l| location.lock().unwrap().push(l));
-            let end = cmp_len;
-
-            let arg_len = location.lock().unwrap().len();
-            match arg_len {
-                2 => {
-                    let _end = location.lock().unwrap()[1];
-                }
-                _ => {}
-            }
-            let start = location.lock().unwrap()[0];
-
-            let feed_start = cmp::max(0, start as isize - start_offset) as usize;
-            let mut feed_end = cmp::min(end + end_offset, *cmp_len);
-            if feed_end - feed_start > 300 {
-                feed_end = feed_start + 300
-            }
-            let mut lines = 0;
-            let mut n_c = 0;
-            for c in composition.clone().chars() {
-                n_c += 1;
-                if n_c > start {
-                    break;
-                }
-
-                if c == '\n' {
-                    lines += 1
-                }
-            }
-            println!(
-                "{}{}",
-                &composition[feed_start..start].yellow(),
-                &composition[start..feed_end].red(),
-            );
-
-            println!(
-                "
-            {}
-            errors at line {}
-            {}
-            ",
-                "working".yellow().underline(),
-                lines.to_string().red().bold(),
-                "broken".red().underline(),
-            );
-
+            handle_parse_error(location, &composition);
             panic!("Unexpected Token")
         }
     }
