+++ songs/test/import_test.socool
@@ -8,6 +8,7 @@ danny = {
 		AsIs,
 		Tm 2,
 	] 
+	| Length 1.0	
 }
 
 main = {
